(c) Microsoft 1988

+++ docker_conf/install_test/docker-run.sh
@@ -5,8 +5,10 @@ if [ "$(id -u)" -ne 0 ]; then
     exit 1
 fi
 
+# default value if no arguments/img are passed
 imgs=("debian" "ubuntu" "fedora" "alpine" "archlinux" "manjarolinux/base")
 dirpath=/docker-data-nobackup/install-test-logs
+LOCAL=$(dirname "$(readlink -f "$0")")
 
 if [ "$#" == 0 ]; then
     echo "No args given testing on ${imgs[*]}"
@@ -17,9 +19,10 @@ fi
 
 function clean() {
     img_name="$(echo "$1" | tr ":/" _)"
-    docker wait cont_"$img_name" &>/dev/null
-    docker kill cont_"$img_name" &>/dev/null
-    docker rm cont_"$img_name" &>/dev/null
+    if ! docker wait cont_"$img_name" &>/dev/null; then
+        docker kill cont_"$img_name" &>/dev/null
+        docker rm cont_"$img_name" &>/dev/null
+    fi
     docker image rm -f "$img_name"_img 1>/dev/null
     echo "$img_name cleaned"
 }
@@ -30,14 +33,16 @@ function create() {
     mkdir -p "$dirpath"
     logpath="$dirpath"/"$img_name"
     rm -f "$logpath/$img_name"
-    tmpD="$(mktemp -d -p .)"
-    cp -fr ../../install.sh "$tmpD"/install.sh
+    tmpD="$(mktemp -d -p /var/tmp/)"
+    cp -fr "$LOCAL/../../install.sh" "$tmpD"/install.sh
+    cp "$LOCAL/dockerfile" "$tmpD"
     printf "#!/bin/bash\n/root/install.sh -n -l /root/%s/logs" "$img_name" >"$tmpD"/test-engine.sh
+    cd "$tmpD" || exit 1
     if docker images | grep "$img_name"_img &>/dev/null; then
         echo "$img_name"_img already created, removing
         docker image rm "$img_name"_img &>/dev/null
     fi
-    docker build --build-arg IMG="$img" --build-arg IMGN="$img_name" --build-arg DIR="$tmpD" -t "$img_name"_img . 1>/dev/null
+    docker build --build-arg IMG="$img" --build-arg IMGN="$img_name" -t "$img_name"_img . 1>/dev/null
 
     docker run \
         -e "TZ=$(timedatectl status | grep "zone" | sed -e 's/^[ ]*Time zone: \(.*\) (.*)$/\1/g')" \
@@ -46,22 +51,20 @@ function create() {
     rm -r "$tmpD"
 }
 
-
-# 3 loops: Img building + run, img and cont cleaning (except logs), logs reading 
-cd "$(dirname "$(readlink -f "$0")")" || exit 1
-echo "Building ${imgs[*]}"
+# 3 loops: Img building + run, img and cont cleaning (except logs), logs reading
+echo "* Building ${imgs[*]}"
 for img1 in "${imgs[@]}"; do
-    create "$img1"
+    create "$img1" &
 done
 wait
 
-echo "Running tests"
+echo "* Running tests"
 for img2 in "${imgs[@]}"; do
     clean "$img2" &
 done
 wait && sleep 2
 
-echo "Results:"
+echo "* Results"
 for img3 in "${imgs[@]}"; do
     img_name="$(echo "$img3" | tr ":/" _)"
     if tail -n 1 "$dirpath"/"$img_name"/logs 2>/dev/null | grep "\[success\] Installation successful" &>/dev/null; then
